[SYCL][CUDA] Event callback worker implementation
Introduces a worker thread living in the platform for handling callbacks
for native events. This is needed as host functions enqueued by
cuLaunchHostFunc is strictly prohibited from doing any CUDA operations,
but as a SYCL object, such as the event itself, may end its life in the
callback we must be able to destroy the CUDA object.

Signed-off-by: Steffen Larsen <steffen.larsen@codeplay.com>